Using .props for library dependency
Using .props for library dependency

+++ Mutator/Custom.h
@@ -14,7 +14,8 @@ namespace TRN
 			std::unique_ptr<Handle> handle;
 
 		public:
-			Custom(const unsigned long &seed, const std::function<void(const unsigned long long &evaluation_id, const unsigned long &seed, const std::vector<int> &offsets, const std::vector<int> &durations)> &request,
+			Custom(const unsigned long &seed, 
+				const std::function<void(const unsigned long &evaluation_id, const unsigned long &seed, const std::vector<int> &offsets, const std::vector<int> &durations)> &request,
 				std::function<void(const unsigned long long &evaluation_id, const std::vector<int> &offsets, const std::vector<int> &durations)> &reply);
 			virtual ~Custom();
 
@@ -22,7 +23,7 @@ namespace TRN
 			virtual void update(const TRN::Core::Message::Payload<TRN::Core::Message::SCHEDULING> &payload) override;
 
 		public :
-			static std::shared_ptr<Custom> create(const unsigned long &seed, const std::function<void(const unsigned long long &evaluation_id, const unsigned long &seed, const std::vector<int> &offsets, const std::vector<int> &durations)> &request, std::function<void(const std::size_t &trial, const std::vector<int> &offsets, const std::vector<int> &durations)> &reply);
+			static std::shared_ptr<Custom> create(const unsigned long &seed, const std::function<void(const unsigned long &evaluation_id, const unsigned long &seed, const std::vector<int> &offsets, const std::vector<int> &durations)> &request, std::function<void(const std::size_t &trial, const std::vector<int> &offsets, const std::vector<int> &durations)> &reply);
 		};
 	};
 };
